Lots of updates here.
So first of all, I changed around the fronted so that there is now
a view page which is immutable, and the edit pages are hidden.

Pages area also now tied to a plaintext url, instead of a hexadecimal
number.

Next, the server was updated to allow session for logged in users. As
well, users cannot edit or delete posts unless logged in. The api has
been updated so that you can check if the current user is logged in.

Lastly, the api uses the url of articles for get requests, but the
hex id for posts, because you can edit the url, and I don't want anything
getting overwritten but mucking about with the url.